updates - add articles status

+++ app/Http/Controllers/ContactController.php
@@ -7,6 +7,14 @@ use Illuminate\Http\Request;
 
 class ContactController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('permission:contacts-create', ['only' => ['create','store']]);
+        $this->middleware('permission:contacts-read',   ['only' => ['show', 'index']]);
+        $this->middleware('permission:contacts-update',   ['only' => ['edit','update']]);
+        $this->middleware('permission:contacts-delete',   ['only' => ['delete']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,9 +22,12 @@ class ContactController extends Controller
      */
     public function index(Request $request)
     {
+        if(!auth()->user()->isAbleTo('contacts-read'))abort(403);
         $contacts =  Contact::where(function($q)use($request){
             if($request->id!=null)
                 $q->where('id',$request->id);
+            if($request->user_id!=null)
+                $q->where('user_id',$request->user_id);
             if($request->q!=null)
                 $q->where('name','LIKE','%'.$request->q.'%')->orWhere('phone','LIKE','%'.$request->q.'%')->orWhere('email','LIKE','%'.$request->q.'%')->orWhere('message','LIKE','%'.$request->q.'%');
         })->orderBy('id','DESC')->paginate();
@@ -31,7 +42,7 @@ class ContactController extends Controller
      */
     public function create()
     {
-        //
+        if(!auth()->user()->isAbleTo('contacts-create'))abort(403);
     }
 
     /**
@@ -42,7 +53,7 @@ class ContactController extends Controller
      */
     public function store(Request $request)
     {
-        //
+        if(!auth()->user()->isAbleTo('contacts-create'))abort(403);
     }
 
     /**
@@ -53,6 +64,7 @@ class ContactController extends Controller
      */
     public function show(Contact $contact)
     {
+        if(!auth()->user()->isAbleTo('contacts-read'))abort(403);
         return view('admin.contacts.show',compact('contact'));
     }
 
@@ -64,7 +76,7 @@ class ContactController extends Controller
      */
     public function edit(Contact $contact)
     {
-        //
+        if(!auth()->user()->isAbleTo('contacts-update'))abort(403);
     }
 
     /**
@@ -76,7 +88,7 @@ class ContactController extends Controller
      */
     public function update(Request $request, Contact $contact)
     {
-        //
+        if(!auth()->user()->isAbleTo('contacts-update'))abort(403);
     }
 
     /**
@@ -87,8 +99,16 @@ class ContactController extends Controller
      */
     public function destroy(Contact $contact)
     {
+        if(!auth()->user()->isAbleTo('contacts-delete'))abort(403);
         $contact->delete();
-        flash()->success('تم حذف طلب التواصل بنجاح','عملية ناجحة');
+        toastr()->success('تم حذف طلب التواصل بنجاح','عملية ناجحة');
         return redirect()->route('admin.contacts.index');
+    }
+
+    public function resolve(Request $request){
+        if(!auth()->user()->isAbleTo('contacts-update'))abort(403);
+        $contact = \App\Models\Contact::where('id',$request->id)->firstOrFail();
+        $contact->update(['status'=>$contact->status=="PENDING"?"DONE":"PENDING"]);
+        return ['status'=>$contact->status=="DONE"?"DONE":"PENDING" ];
     }
 }
